Moved code generation classes to the Generation folder
