+++ src/Abioc/Generation/CodeGen.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) 2017 James Skimming. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-namespace Abioc.Composition
+namespace Abioc.Generation
 {
     using System;
     using System.Collections.Generic;
